pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "otto"
version = "0.1.0"
description = "CLI tool for monitoring and orchestrating containers"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "lolosioann", email = "lolosioann@gmail.com" }
]
dependencies = [
    "docker>=7.0.0",
    "aiodocker>=0.21.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "commlib-py>=0.11.0",
    "psutil>=5.9.0",
    "structlog>=23.0.0",
    "watchdog>=3.0.0",
    "paho-mqtt>=2.1.0",
    "pre-commit>=4.5.1",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "nox",
    "ruff",
    "bandit",
    "interrogate",
    "sphinx",
    "sphinx-rtd-theme",
    "myst-parser",
]

# ------------------------
# Ruff configuration
# ------------------------
[tool.ruff]
line-length = 100
target-version = "py310"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM", "C4"]
ignore = ["E501"]
per-file-ignores = { "tests/**" = ["S101"] }

[tool.ruff.format]
quote-style = "double"

# ------------------------
# Pytest configuration
# ------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--strict-markers --strict-config"
pythonpath = ["."]

# ------------------------
# Coverage configuration
# ------------------------
[tool.coverage.run]
branch = true
source = ["src"]
omit = ["*/__init__.py", "*/cli/*"]

[tool.coverage.report]
fail_under = 85
show_missing = true

# ------------------------
# Bandit security scanner
# ------------------------
[tool.bandit]
skips = ["B101"]
exclude_dirs = ["tests"]

# ------------------------
# Interrogate docstring enforcement
# ------------------------
[tool.interrogate]
ignore-init-method = true
ignore-init-module = true
ignore-module = false
ignore-private = true
ignore-magic = true
fail-under = 80
exclude = ["tests", "test_scripts"]

[dependency-groups]
dev = [
    "myst-parser>=4.0.1",
    "pyenchant>=3.3.0",
    "sphinxcontrib-spelling>=8.0.2",
]
